refactor(tui): remove legacy composer and header
